+++ crates/bitwarden/Cargo.toml
@@ -59,6 +59,7 @@ subtle = ">=2.5.0, <3.0"
 thiserror = ">=1.0.40, <2.0"
 uniffi = { version = "=0.25.2", optional = true, features = ["tokio"] }
 uuid = { version = ">=1.3.3, <2.0", features = ["serde"] }
+zxcvbn = "2.2.2"
 
 [dev-dependencies]
 rand_chacha = "0.3.1"
