+++ crates/bitwarden/src/auth/password/mod.rs
@@ -1,0 +1,7 @@
+mod policy;
+pub(crate) use policy::satisfies_policy;
+pub use policy::MasterPasswordPolicyOptions;
+mod validate;
+pub(crate) use validate::validate_password;
+mod strength;
+pub(crate) use strength::password_strength;
